locality/server/requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==37.0.4
Django==4.0.6
django-cors-headers==3.13.0
django-rest-auth==0.9.5
django-rest-knox==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
mysqlclient==2.1.1
pycparser==2.21
PyJWT==2.3.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
